+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ja.dart
@@ -15,6 +15,9 @@ class StreamChatLocalizationsJa extends GlobalStreamChatLocalizations {
   String get noUsersLabel => '現在、ユーザーはいません。';
 
   @override
+  String get noPhotoOrVideoLabel => '写真やビデオはありません';
+
+  @override
   String get retryLabel => '再試行';
 
   @override