+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_pt.dart
@@ -15,6 +15,9 @@ class StreamChatLocalizationsPt extends GlobalStreamChatLocalizations {
   String get noUsersLabel => 'Nenhum usuário atualmente';
 
   @override
+  String get noPhotoOrVideoLabel => 'Não há foto ou vídeo';
+
+  @override
   String get retryLabel => 'Tente novamente';
 
   @override